requirements.txt
pycrypto==2.6.1
Cython==0.29.23
pyinstaller==4.4